libstorage: ext4: Fix fsck error handling
Fix logic issues when we need to recreate the filesystem.
fsck will try to qualify errors (FSCK_ERROR bit set); when it can't
it means the filesystem is in really bad shape.
Reformat it since we won't be able to mount it.

BUG=b:376103371
TEST=Install dm-default-key layout, which reformat stateful.
Check Setup is called twice without issue.

Change-Id: Ifb2b47a3ea8f2c855b91ee13393e1065bd6e2dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6093198
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>