modemfwd: make Modem reference counted and synchronize access
This object is used across threads, though it really should
only be used by one thread at a time. However, it is possible
that multiple threads can try to use it at the same time, so
we should make sure that's not happening. It is also possible
that this could lead to use-after-free problems as the
ownership semantics are more complicated now.

BUG=b:362262152,b:362263331,b:378152823
TEST=unit tests, deploy and force-flash on rusty

Change-Id: If25a016ad4c7dd70e8ff2ecf7bc212beed79eec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6075512
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Rukun Mao <rmao@google.com>