cryptohome: Get rid of explicit template instantiations
These are used in a few places, but they're generally a bad idea as they
can be fairly brittle and produce bad errors (even moreso than usual for
templates) when you're missing an instantiation.

In one case the implementations are trivial and so are fairly easy to
just pull into the header. In the other we can use forwarding overloads,
which requires a bit more boilerplate but which is generally more robust
to failure.

BUG=none
TEST=unit tests

Change-Id: I022a2c6b38dee80b6b3893c704dc23a6284eda1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6076054
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>