cryptohome: Get rid of most of the remaining "-" names
This is a followup to a prior change that got rid of most of the
*_unittest.cc names, this also gets rid of most of the uses of dash in
code filenames. This makes everything more consistent with the general
style, and the rest of the code.

Note that this does NOT rename any of the executables that get built.
There's no need for the binary names to match the source code, so it's
okay to have stuff like a lockbox-cache binary but all of the source
code uses lockbox_cache (and this is in fact common in Google code).

BUG=none
TEST=unit tests, CQ

Change-Id: Iaf3a8e3c2541f9a4d11518485a0bb6b8f443b5e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6085757
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>