vm_tools: Simplify StartVhostUserFs() API
Previously, the user of vhost_user_client is responsible to format a
vector with only 1 ScopedFd as the argument of StartVhostUserFs()
function. This is because StartVhostUserFs dbus method uses a vector of
fd as in argument. However, user does not to need to know the detail.

This CL makes StartVhostUserFs() take a ScopedFd instead of a vector of
ScopedFd as parameter. The vector of ScopedFd will be created inside the
StartVhostUserFs() function.

BUG=b:355159487
TEST=CQ

Change-Id: I80c0f7f824990760385398f2955f398876160080
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6086763
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Yuan Yao <yuanyaogoog@chromium.org>
Tested-by: Yuan Yao <yuanyaogoog@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>