debugd: Fix error management in capture_utility.sh
`configure_monitor` function on capture_utility.sh directly emits the
error message into stdout and it propagates as a wrong network device in
the error method. Capture the error into a variable to print correct
error message.

Fix the error with the float value as a frequency in configure_monitor
which causes the attached bug.

BUG=b:381364601
TEST=Run cros emerge-{$BOARD} debugd and run "packet_capture
--monitor-connection-on wlan0 --frequency 5280" on DUT.

Change-Id: If2ff386f25bae0df757fbfce2f619f8900972d83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6088409
Commit-Queue: Irem Uguz <iremuguz@google.com>
Tested-by: Irem Uguz <iremuguz@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>