cros-codecs: Fill out vp8 header
All of the frame information sans reference frame management
is filled out. This would hopefully allow for key frame only
clips to be played. Unfortunately this is not the case, when
playing a stall occurs with buffers awaiting to be dequeued.

BUG=b:382563536
TEST=Attempt to decode vp8 clip on asurada

Change-Id: I7270844e6cc3dc18954f972e0bf2b5998e674f42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6097210
Tested-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>