ml_core: Disable OpenCL caching if OpenCL is not available
On platforms like rauru, we don't enable CLVK. Thus, we cannot use
OpenCL as the delegate, which means we don't need to cache its output.

This CL adds a check of the existence of the OpenCL library to decide
whether to enable OpenCL caching.

BUG=b:379062550, b:378633716
TEST=CQ

Change-Id: I5975163fcf8c46d9fbe89f672033e92be636c539
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6100577
Reviewed-by: Imran Ziad <imranziad@chromium.org>
Tested-by: Tommy Chiang <ototot@google.com>
Commit-Queue: Tommy Chiang <ototot@google.com>