cros-codecs: Fix AV1 frame size override logic
When the frame size override flag is set to true, the parser should read
the `found_ref` bits to find a reference frame to set the frame size to.
Currently this bit is interpretted to correspond directly with a
reference frame, but it should actually correspond to an entry in the
`ref_frame_idx` table, which in turn corresponds to a reference frame.

BUG=b:382695401
TEST=video.PlatformDecoding.cros_codecs_vaapi_av1_files_from_bugs_346405213

Change-Id: I128c6207c066da856ae81793da54de4e9d5ac210
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6096312
Commit-Queue: Justin Green <greenjustin@google.com>
Tested-by: Justin Green <greenjustin@google.com>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>