Shill: WiFi: Assign Correct Disconnect Type for Handshake Timeouts
Currently, when a WiFi connection fails due to a handshake timeout, the
disconnect type is incorrectly labeled as kWiFiDisconnectTypeShill. This
is because WiFiService::OnDisconnect doesn't explicitly set the
disconnect type for this scenario.

Assign the appropriate disconnect type
kWiFiDisconnectTypeHandshakeTimeout when a handshake timeout occurs.

BUG=b:363102970
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ib478b928bd30301a55547a111a637efc443386ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6102594
Tested-by: Kaidong Wang <kaidong@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kaidong Wang <kaidong@chromium.org>