data_migrator: Skip migration of fs-verity enabled files
This CL adds an option to MigrationHelper to skip the migration of files
that enable fs-verity if opening the source file fails, and enable the
option for ARCVM virtio-blk /data migration (but disable for the
dircrypto migration to preserve its behavior for now). In ARCVM
virtio-blk /data migration, attempting to open a fs-verity file in the
migration source as writable always fails, so we should skip its
migration. This CL also extends the enum used for the
Arc.VmDataMigration.FailedPathType UMA to record the occurrence of this
skip logic on the problematic font file.

BUG=b:384841851
TEST=`tast run <DUT> arc.VMDataMigration` with CL:6109241 passes.

Disallow-Recycled-Builds: test-failures
Change-Id: I683f0a3b09d1a9da843be125d8d95a34ecdc8854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6109592
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>