cryptohome: Add some assertion enhancements the iterators
This also adds a default constructor, which is technically required for
forward iterators (and the assertion helpfully catches). It's just a
default-valued one that would be undefined to use for anything other
than assigning over it, but fortunately that's all that's expected of
such iterators.

BUG=None
TEST=unit tests

Change-Id: I65a2b62308f7126ceeb3d1b832e0f18c3ffa53d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6110268
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Commit-Queue: Zi Lin <lziest@google.com>