installer: Rename RunCommand to Platform
This is in preparation for adding more mockable operations, such as
`unmount`. Calling it "Platform" matches C++ platform2 code with similar
abstractions. See for example cros-disks/platform.h and
libstorage/platform/platform.h.

Also rename the get_output_as_string method to
run_command_and_get_stdout, and clarify in the docstring that stderr is
captured as well (but not returned to the caller).

BUG=b:384912299
TEST=FEATURES=test emerge-reven chromeos-installer

Change-Id: Ib2ac1c045a19344f66ecc68c82c6264fe24ba219
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6108022
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>