patchpanel: dbus: Adjust some log severity
In OnSocketAnnotation(), as suggested in the existing comment, it's
possible that this function is called when the client is already
destroyed (e.g., during shutdown). It's usually not something we need to
worry about, let's use WARNING here instead.

Also note that  in the case of base::PostTask(), when the object which
the task is bound to already out of scope, the default behavior is
discard the task silencely, so in general, we are not caring about such
cases that much.

BUG=b:374658759
TEST=None

Change-Id: I7dddae4df477fcb75580f52ab00d699f3f306e77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6124375
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>