installer: Return dd and losetup tricks to shell
We moved the tricks to find the right version of dd and losetup from
shell to Rust, but it turns out it doesn't work quite like we expected.
On both MiniOS and Flexor it seems that dd can't be called directly
from Rust because it's actually an alias for busybox. Also, at least on
Flexor, Rust seems to find the non-busybox (correct) losetup by default.

By the time we were alerted to the issue a large number of changes had
landed on top, and I'm not sure about the best way to handle that
through gerrit. A pure revert here is harder than manually undoing
these changes, so let's just manually undo them.

BUG=b:384948237
TEST=cargo test
TEST=Running a flexor_disk.img created with `create_flexor_disk` and
     a flexor_vmlinuz containing this version of chromeos-installer
     results in a successful install.
TEST=cargo xtask test-install chromiumos_test_image.bin

Change-Id: Ibdb7431f90d41580c56efc320292286c0a8728c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6127331
Commit-Queue: Saketh Pothireddy <spothire@google.com>
Auto-Submit: Ted Brandston <tbrandston@google.com>
Reviewed-by: Saketh Pothireddy <spothire@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>