odml: Add suppression_context for coral
The suppression_context option in GroupRequest would allow the caller to
indicate to odmld to that it does not wish to have certain contents
grouped. Basically, if a specific proportion of entities in
suppresion_context is found in a group, then it will be suppressed.

BUG=b:388004243
TEST=unit test

Change-Id: I978c621780341cabb247cf0b67bfb4df67351b1f

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6149468
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>