libmems: Move static functions and GetContext to base IioDevice class
iio_device static functions (to process iio:device string) can be
in the virtual iio_device object, they don't need to be in the real iio
device implementation object.

Also the iio_context the iio_device belongs to can also be in the
virtual class.
The test iio_context class is created and manipulated outside of the
iio_device object, so it is fine if GetContext() return the virtual
iio_context class, not the test one.

BUG=none
TEST=Unit test.

Change-Id: Ie87d90d842e0e5dba52a0b8121e0644f2e256f22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6164340
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Harvey Yang <chenghaoyang@chromium.org>