runtime_probe: Parse return value of i2c read as bytes for ec_component
ec_component casts the return value of i2c read to int, and do
comparison with `expect` in int. The casting is endianness-sensitive.

Since ec_component will support probing multiple bytes, and the values
in component manifests are endianness-insensitive, we need an
endianness-insensitive way to do the comparison. This CL modifies
ec_component and ec_component_manifest to parse the values in bytes
without casting to int.

BUG=b:380054141
TEST=EATURES=test emerge-geralt runtime_probe

Change-Id: I735d13b7fa88debfebc77e36f275b000b9ddd749
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6164546
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>