cryptohome: Fix segfault in auth session expiration code
These changes should address a low-frequency segfault we're seeing
coming from the auth session expiration code. The failure seems to be an
assumption during a lookup of the AuthSession object that the session
will never be in use, which fails if it is. This reworks the code to not
depend on finding the AuthSession object at all in order to look up the
public token for the session.

While having a private->public token map isn't ideal, this isn't really
any worse than all the copies of the private token the session manager
already has; it kind of needs them since it's what manages the session.
In practice it's not any worse than the already existing private token
-> obfuscated user map.

BUG=b:389154305
TEST=unit tests

Change-Id: I5905054b308b16822c8f2c5d92f089895b062f33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6169221
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>