flexor: Log output from child processes
Flexor runs several child processes (including chromeos-install). The
output (stdout and stderr) from these processes can be helpful for
debugging. Prior to this commit the output didn't go anywhere
useful. Now it is sent to the logger, which sends it to /dev/kmsg.

Note that in order for this output to be visible the kernel loglevel
must be at least 5, and printk.devkmsg=on must be set to turn off rate
limiting.

This code is adapted from platform2/os_install_service.

BUG=b:389708383
TEST=FEATURES=test emerge-reven chromeos-base/flexor
TEST=Build+run flexor image, observe chromeos-install logging

Change-Id: Ia46fe2d054e038415817f7a78bb51fc086fd5b19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6173252
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>