vm_tools/maitred: Use SetUserLinger D-Bus API to enable linger
Earlier we were manually creating a linger file for our new user in
Baguette. It appears that systemd-logind needs to be restarted for
it to pick up this manually created linger file.

In this CL, we instead issue a `SetUserLinger` message on the
system D-Bus. This also avoids us having to hard code the linger
path for systemd.

BUG=b:388544890
TEST=Deploy termina-dlc, \
     destroy and recreate a Baguette VM, \
     check `loginctl list-users`

Change-Id: I34d88c9dbc54a12cdc1fdc18df68f79257fc344c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6160434
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Auto-Submit: Saswat Padhi <spadhi@chromium.org>
Commit-Queue: Saswat Padhi <spadhi@chromium.org>
Tested-by: Saswat Padhi <spadhi@chromium.org>