cros-codecs: Rework queuing of v4l2 buffers
The allocation of buffers was based on returning early when
the decoder ran out of OUTPUT buffers. This required there to
be the same number of CAPTURE and OUTPUT buffers.

Because the hardware is not pipelined it can only use one buffer
at a time. While one buffer is decoding, the next can be filled.
An optimal decoder will only need 2 OUTPUT buffers and ideally
dpb CAPTURE buffers.

There is still a logic problem with the dpb and a magic number
is used to make the buffers flow through.

BUG=b:382293969
TEST=test-25fps.h264

Change-Id: I735d1a232b6435f524c39a420258d568efbe8fd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6180741
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>
Tested-by: Fritz Koenig <frkoenig@chromium.org>