missive: Timers to validate weak pointers
After discussion with run time library owners, it appears that timers
do not skip running scheduled callbacks (as opposed to PostTask) if
their weak pointers are already invalidated, and thus they can crash.

To prevent this, changing all timers callbacks to explicitly validate
the weak pointers of their base classes and bail out if nullified.
This way even if the callback is already scheduled but not run before
the object is destructed, the timers will not crash.

BUG=b:385965939
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual on VM

Change-Id: Ic1f591eff965990dee1125612078716527b27d15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6180341
Reviewed-by: Josh Hilke <jrhilke@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>