fpstudy: Add Experiment parsing from decisions dir
This allows for creating an encapsulating TestCase with
a TestCaseDescriptor and Experiment.

This further aids the mission to adapt TestCase to contain
an Experiment, that can be parsed from a directory.
This is in alignment with https://crrev.com/c/5664058,
where a TestCase should contain an Experiment.

BUG=b:388931275
TEST=cros format .
TEST=cros lint .
TEST=python3 -m unittest discover -v -p '*_test.py'
TEST=./simulate_fpstudy.py simulation
     ./run.py analyze simulation
     rm simulation/FAR_decisions.csv
     ./run.py analyze simulation
     # ERROR CURRENTLY, DUE TO unique_list depending on FAR TABLE

Change-Id: Iec13495f6c7682b653b8b8d8d781a5966bfb8c6b

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6184152
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>