login: Move BrowserJob creation to SessionManagerService.
BrowserJob real implementation will depend on a member of
SessionManagerService in a following CL, so this is the preparation
for the change.

Along with the change, ShouldRunBrowser() is moved out from
BrowserJob to SessionManagerService, because the service (not a job)
controls whether or not to restart Chrome process.

BUG=b:390297821
TEST=Tryjob

Change-Id: I2821b5806b6e18cb080768575c3e0ccc5546f070
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6187041
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>