printscanmgr: make daemon broker inside isolated network
This change is part of the project of bringing ipcz to ChromeOS.

The printscanmgr service uses Mojo to communicate between its
daemon and its executor. However, it doesn't use Mojo to
communicate with the rest of the system.

Every Mojo network needs to have a broker, and in this case,
it can not be inherited from the existing network.

This change makes it so that the daemon takes the role of
broker in the printscanmgr daemon/executor network, which
is required for ipcz to work.

BUG=b:41494930
TEST=tast run ${DUT} printer.PrintscanDebug on an image built
with USE=ipcz

Change-Id: Iebffa8e16e48142b3ed214a5cc123b80da4abc26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6189138
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Tested-by: Andrea Orru <andreaorru@chromium.org>