camera: diagnostics: allow access to /dev/shm and few syscalls for ipcz
As part of the effort of bringing ipcz to ChromeOS, we need to
enable a few additional syscalls for cros-camera-diagnostics.

Additionally, we need to grant write access to /dev/shm to grant
cros-camera-diagnostics the ability to initiate shared-memory-based
communication via MojoIpcz.

BUG=b:41494930
TEST=tast run ${DUT} health.RunCameraFrameAnalysisRoutine
on an image built with USE=ipcz

Change-Id: I8d94abd4a6101f21c8942c1eed3f8f47c8d02c15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6192296
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Andrea Orru <andreaorru@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>