cros-toolchain: clean up use of global
It was highlighted in crrev.com/c/6180339 that globals are kind of
cringe if they can be avoided. I didn't want to address that feedback in
the CL, since that was meant to be a minimal-changes-copy-and-paste of
sources from an old location to a new one. Nothing stopping me from
addressing it now, though.

BUG=b:388925046
TEST=FEATURES=test emerge-arm64-generic toolchain-tests

Change-Id: I2f5c81c43cb89864340e3d6a9e86a2b39f241e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6192943
Reviewed-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>