power: Add new backlight cause for battery saver state changed
This CL fixes the linked bug by adding a specific request type for when
the battery saver state changes. On the Chrome side, we save the user's
preferred brightness if the change cause was USER_REQUEST. By making a
separate type for when battery saver state changes, we eliminate that
case and thus the source of the bug.

Manual test steps:
1. Plug in DUT to AC power, charge until mostly full
2. Set brightness to max
3. Unplug AC power
4. Observe that brightness stays bright (not dim)
5. Enable Battery Saver mode
6. Observe that brightness dims to 20%
7. Power off DUT. When it’s off, plug in AC power.
8. Power on DUT
9a. BEFORE (bug): Brightness is still 20%.
9b. AFTER (fix): Brightness is back to full (or previous user setting)

BUG=b:387207153
TEST=internal_backlight_controller_test + manual test
Change-Id: I4a77478c9db5a8e21d1d83b1b65a0439d6d05623
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6194669
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Tested-by: Camden Bickel <cambickel@google.com>
Commit-Queue: Camden Bickel <cambickel@google.com>