vm_tools: concierge: add source-image based disk creation.
As part of baguette installation, we want to use the provided
zstd-compressed disk image as the writable rootfs. This patch adds the
ability for concierge to use such an image, decompress it into the disk
image, and then proceed as normal.

BUG=b:377377749
TEST=build dev-rust/system_api chromeos-base/system_api vm_host_tools crostini_client
TEST=vmc create --size 20G --source baguette.img.zstd
TEST=vmc start .... baguette (see https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform2/vm_tools/baguette_image/README.md)

Cq-Depend: 6207802
Change-Id: I0c1903bbd16f089a1b44388eeef5e2333bc2cb52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6208610
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: maciek swiech <drmasquatch@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Zihan Chen <zihanchen@google.com>