cryptohome: Replace a bunch of std::map with flat_hash_map
The ordering that std::map provides should not be relevant in any of
these cases, so flat_hash_map should be a boost in efficiency.

One caveat is that there were some tests which depended on the output
order of some of the code in question. I believe this was accidental and
not deliberate, and so I just fixed the tests, but I could be wrong.

BUG=None
TEST=unit tests

Change-Id: I6e3902d8135b73b23025b5a397faaf0a1cf10811
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6217080
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>