flexor: Make update-test-disk faster
The code to update the disk image was based on crdyboot's xtask, which
only operates on the ESP. The ESP is usually fairly small, but the
flexor data partition is large (in a typical image, over 30GiB). Loading
it into memory, modifying it, then writing it back out is quite
slow. Speed it up by directly reading and writing to the file on disk.

This improves an update operation from about 45s-75s to about 4s.

To make this possible, add a `FileView` type that acts like a regular
File, but operates only on a slice within the file.

Testing:
* Generate a flexor image that doesn't have any files in the flexor data
  partition.
* Run with `cargo xtask run-test-disk <path>` and verify that crdyboot
  fails to launch flexor.
* Run `cargo xtask update-test-disk --flexor-vmlinuz <path>
  --install-image <path> <output>`.
* Run with `cargo xtask run-test-disk <path>` and verify that flexor
  succeeds now.

BUG=b:389940448
TEST=As described above
TEST=cargo test -p xtask

Change-Id: I6a8a1dcec043ecdf59577ba65d0575bd1e1745a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6221996
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Jiao Zhou <jiaozhou@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>