ippusb_bridge: Convert from tiny_http to hyper
ippusb_bridge currently uses a forked copy of an older version of
tiny_http for processing incoming HTTP requests.  In order to stop
maintaining this fork, convert to use the standard hyper crate instead.

Unlike tiny_http, hyper is asynchronous and depends on the tokio
runtime.  tokio was already introduced in crrev.com/c/6279033, so
ippusb_bridge had already been partially converted to async processing.
This CL extends that to cover the main HTTP responder in addition to the
internal unix socket forwarder.

On the USB side of the connection, rusb/libusb are not asynchronous.  On
the TCP listening side, the epoll wrappers are also synchronous.
Because hyper in the middle is now asynchronous, the result is a
sandwich of sync <-> async <-> sync functions, but this is hidden away
inside the Daemon::run function.  For external printing and scanning
callers, there should be no behavior differences except for possible
timing and HTTP body chunk sizes.

Since this is a large change that doesn't have other obvious external
indicators, also update the version 0.2.0 to make it easier to tell
right away if we get reports about changes in behavior.

BUG=b:393602658
TEST=tast run '("group:paper-io")'; manual USB printing and scanning

Change-Id: I7cd7fe22ad5d15350c73e98c76a2f81f4f55070a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6314003
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>