pmt_tool: initial commit for external netdata plugin
Initial commit for an external netdata plugin that displays Intel PMT data.
Using Netdata External Plugin API:
https://learn.netdata.cloud/docs/developer-and-contributor-corner/external-plugins
If PMT data is not available in a csv file, it starts sampling.
To allow netdata to call pmt_tool for sampling, netdata has to be run as root.

By default, the plugin is marked disabled in the netdata.conf file.
To turn the plugin on, the user must manually enable it by editing the conf file.

BUG=b:354134139
TEST=emerge-rex chromeos-base/pmt_tool
TEST=edit netdata.conf; intel_pmt = yes
TEST=netdata -u root OR start_netdata.sh

Tested-by: Daniel Balint <dbalint@google.com>
Change-Id: I57abcc35a428c2a11a7fac5f68a92357b30dc58b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6225744
Reviewed-by: Stanisław Kardach <skardach@google.com>
Commit-Queue: Marcin Wojtas <mwojtas@google.com>
Reviewed-by: Marcin Wojtas <mwojtas@google.com>