midis: remove unsupported, unnecessary teardown operations for ipcz
This change is part of the project of bringing ipcz to ChromeOS.

MojoIpcz does not require explicit shutdown operations and does
not provide the equivalent of "GetActiveHandlesForTest".

This code was originally added here: https://crrev.com/c/1970104
to mimic legacy EDK shutdown, but it's currently the only instance
of this usage in the whole Chrome/ChromeOS codebase.

BUG=b:41494930
TEST=USE="ipcz" cros_workon_make --board=${BOARD} chromeos-base/midis --test

Change-Id: I997b842815855d5162f72d20cbf02cb1aedd0f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6232670
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Andrea Orru <andreaorru@chromium.org>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>