camera: fix IsUsbCamera
The vid_pid field can be just ":" for a MIPI camera device, which is not
empty. This led to a crash on a platform with both MIPI and USB cameras
if upstream IPU6 ISYS driver is used. Fix this, by using bcd_device
instead to determine a USB video node.

BUG=b:394536206
TEST=media_v4l2_test --list_builtin_usbcam does not crash
TEST=tast run camera.KernelSmokeUSB

Change-Id: I71a31891e0d382aa7ec69547877c5a3b53ba8c9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6241796
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Hidenori Kobayashi <hidenorik@chromium.org>
Tested-by: Hidenori Kobayashi <hidenorik@chromium.org>