cros-codecs: Implement PooledVideoFrame
Implement a variant of VideoFrame which is compatible with FramePools.
PooledVideoFrame wraps an existing VideoFrame type and forwards all of
the core functionality to the inner frame, but its Drop implementation
returns it to the FramePool (if there hasn't been a DRC) so that it can
be re-used.

FramePools are necessary not only to avoid expensive DMA buffer
allocations, but to cap the number of File Descriptors in use at any
given time. Without FramePools, some test vectors will cause ccdec to
allocate a very large number of DMA buffers and exhaust our FD limit.

BUG=b:390439314
TEST=video.PlatformDecoding.cros_codecs* on Brya and Corsola

Change-Id: I68121501beb2a24d93c37dc1ad71228970895cc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6254220
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Justin Green <greenjustin@google.com>
Tested-by: Justin Green <greenjustin@google.com>