shill: portal: upgrade CAPPORT HTTP URLs to HTTPS
The CAPPORT RFC8908 specifies that connections to the portal sign-in
page should be over TLS, therefore the user-portal-url field of the
CAPPORT JSON answer is expected to be HTTPS. However, Android happens to
allow HTTP URLs and CAPPORT impls returning HTTP URLs for
user-portal-url have been found in the wild.

This patch relaxes the requirements around user-portal-url and
automatically update the scheme to HTTPS when an HTTP user-portal-url is
found.

BUG=b:396556880
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ieb7e884b30bb121cc8515d502001bf2a2e9510cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6267743
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>