cros-codecs: Handle VP9 superframes
Superframes contain a non-shown frame. Reference frame
management needs to take into account that the frame is
not shown, but used as a reference.

To track them they need to be associated with a video frame.
When they are dequeued from the driver the decompressed
buffers need to be associated with video frame. Previously they
were not associated the buffers and would stay in the hash map.
This provided a reference that prevented v4l2r from advertising
the buffer as usable. Every superframe would eat a v4l2
buffer.

The buffers are now associated with a frame as soon as they
are dequeued. There was already a hash map holding the request
object. The hash maps can be consolidated into one that
drops the references and associates the buffer.

BUG=b:384980627
TEST=ccdec test-25fps.vp9 --input-format vp9

Change-Id: I25d027e941e95623444654e95838721b65254f94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6282705
Tested-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>