shill: cellular: Add fallback attach APN when no attach exist
For the revamp UI logic, add a fallback attach APN when the user has a
custom APN list but there is no attach APN in the list. The fallback
APN is needed to force modem manager to update the initial EPS bearer
settings in the modem, so the modem can trigger an attach proceduce
when the automatic attach is disabled in the modem.

BUG=b:386319482
TEST=On kracko_fm101_cat12, use test_no_apns.pbf so no MODB APNs are
used. On the UI, enter a default APN without attach APN.
Restart modem. Verify that the attach try list is empty.
Install this patch, and follow the same procedure.
Now the try list includes the attach APN, and the force flag is set to
true.

Change-Id: I86e4a3650558af2c6ad05bc4b8c34c7cee595855
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6271965
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Rukun Mao <rmao@google.com>