modemfwd: Force the purge of all unused DLCs on boot
A while back, the behaviour of GetExistingDlcs in dlcservice was
modified to only return the list of Installed DLCs. Previously, all DLCs
were returned, even the ones that were not installed if they contained
preloaded files in the DUT. This is not an issue for user devices, but
on lab devices, the preloaded DLCs remain, because modemfwd does not
delete them. Instead of purging the DLCs based on GetExistingDlcs,
purge all unused DLCs that are listed in the manifest.

BUG=b:397417828
TEST=tast run lab_anraggar_fm101_att cellular.ModemfwdPurgeDlcs

Change-Id: I93dc3100e38127a61ae8de823b42951c6c73c945
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6279506
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>