ippusb_bridge: Refactor socket handling
ippusb_bridge can listen on a unix socket or a TCP socket.  As a first
step to migrating to TCP only, change this to always listen on a TCP
socket.  For backwards compatibility, also start up a unix socket that
forwards to the TCP port if the socket path is passed in.

This uses tokio for the forwarding path.  This would normally be an
unreasonably large dependency for a modest feature like this even though
it makes the code simple.  However, we are planning to migrate more of
ippusb_bridge to use tokio anyway, so we might as well introduce it
here.

BUG=b:393602658
TEST=tast run '("group:paper-io")'; manual printing

Cq-Depend: chromium:6279301
Change-Id: I0d7d799463900b6af49d5cad44d5ed6fdb755016
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6279033
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>