cros-codecs: Implement unaligned MM21 detile logic
Prior to this CL, the MM21 to NV12 detiling logic would assume that the
frames are aligned with the MM21 tile width. This same logic does not
work if the frame is not aligned.

Updates the detiling logic to support frames that are not aligned with
the MM21 tile width. This logic is translated from the same logic in
libyuv.

BUG=b:392174530
TEST=N/A

Change-Id: I190662af4bbc64a0a010ee74086f05f3bb4f2d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6282706
Reviewed-by: Justin Green <greenjustin@google.com>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Tested-by: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>