libchromeos-rs: Add Efivarfs type for mounting
Needed by flexor to mount Efivarfs.

Worth noting that the ergonomics here are imperfect: you still need
to pass a path to the Builder, even though it doesn't mean anything.
This matches the actual behavior of mount (it's recommended you pass
`none` as the source when mounting efivarfs), but we might be able to
do better. Follow up bug: b/398837589.

BUG=b:388506100
TEST=CARGO_TARGET_X86_64_UNKNOWN_LINUX_GNU_RUNNER='sudo -E' \
     cargo test -- --ignored

Change-Id: I6ec8ae7d2235792762297db39c16d8dc1e49aedb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/6298678
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>